+++ test/integration/checkPage.js
@@ -25,6 +25,11 @@ describe('Check page', () => {
     expect($('.button__check').prop('href')).to.equal(`${constants.siteRoot}${routes.search.path}`);
   });
 
+  it('has an urgent help call out', () => {
+    expect($('.samaritans__call').text()).to.equal('116 123');
+    expect($('.samaritans__email').text()).to.equal('jo@samaritans.org');
+  });
+
   it('has a back link to the start page', () => {
     iExpect.backLinkContent($, `${constants.siteRoot}/`);
   });